handle reservation metadata
