+++ resources/api/reservation.py
@@ -36,7 +36,7 @@ from resources.models import (
     Reservation, Resource, ReservationMetadataSet,
     ReservationHomeMunicipalityField, ReservationBulk, Unit
 )
-from resources.models.reservation import RESERVATION_BILLING_FIELDS, RESERVATION_EXTRA_FIELDS
+from resources.models.reservation import RESERVATION_EXTRA_FIELDS
 from resources.models.utils import build_reservations_ical_file
 from resources.pagination import ReservationPagination
 from resources.models.utils import generate_reservation_xlsx, get_object_or_none
@@ -44,12 +44,11 @@ from resources.models.utils import generate_reservation_xlsx, get_object_or_none
 from ..auth import is_general_admin, is_underage, is_overage, is_authenticated_user, is_any_admin, is_any_manager
 from .base import (
     NullableDateTimeField, TranslatedModelSerializer, register_view, DRFFilterBooleanWidget,
-    ExtraDataMixin
+    ExtraDataMixin, ReservationCreateMixin
 )
 
 from ..models.utils import has_reservation_data_changed, is_reservation_metadata_or_times_different
 from respa.renderers import ResourcesBrowsableAPIRenderer
-from payments.utils import is_free, get_price
 
 from maintenance.models import MaintenanceMode
 
@@ -87,83 +86,6 @@ class UserSerializer(TranslatedModelSerializer):
         fields = ('id', 'display_name', 'email')
         ref_name = 'ReservationUserSerializer'
 
-class ReservationStackSerializer(serializers.Serializer):
-    begin = NullableDateTimeField()
-    end = NullableDateTimeField()
-
-class ReservationBulkSerializer(serializers.Serializer):
-    preferred_language = serializers.CharField(required=False)
-    reserver_email_address = serializers.EmailField(required=False)
-    reserver_name = serializers.CharField(required=False)
-    reserver_phone_number = serializers.CharField(required=False)
-    resource = serializers.PrimaryKeyRelatedField(queryset=Resource.objects.all())
-    type = serializers.ChoiceField(choices=Reservation.TYPE_CHOICES)
-    reservation_stack = ReservationStackSerializer(many=True)
-    comments = serializers.CharField(required=False)
-
-    class Meta:
-        fields = (
-            'preferred_language', 'reserver_email_address',
-            'reserver_name', 'reserver_phone_number',
-            'resource', 'type', 'reservation_stack',
-            'comments'
-        )
-
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        data = self.get_initial()
-        resource = Resource.objects.get(pk=data['resource'])
-        cache = self.context.get('reservation_metadata_set_cache', None)
-        required = resource.get_required_reservation_extra_field_names(cache=cache)
-        for field_name in required:
-            self.fields[field_name].required = True
-    
-    def validate(self, attrs):
-        attrs = super().validate(attrs)
-
-        resource = attrs['resource']
-        _cattrs = attrs.copy()
-        reservation_stack = _cattrs.pop('reservation_stack')
-        for data in reservation_stack:
-            reservation = Reservation(**_cattrs, **data)
-            reservation.clean()
-            if resource.validate_reservation_period(reservation, reservation.user):
-                return JsonResponse({
-                    'status': 'false',
-                    'recurring_validation_error': _('Reservation failed. Make sure reservation period is correct.')
-                }, status=400)
-            if resource.validate_max_reservations_per_user(reservation.user):
-                return JsonResponse({
-                    'status': 'false',
-                    'recurring_validation_error': _('Reservation failed. Too many reservations at once.')
-                }, status=400)
-
-        return attrs
-    
-    def validate_reserver_phone_number(self, value):
-        if value.startswith('+'):
-            if not region_code_for_country_code(phonenumbers.parse(value).country_code):
-                raise ValidationError(dict(reserver_phone_number=_('Invalid country code')))
-        return value
-
-    def create(self, validated_data):
-        reservation_stack = validated_data.pop('reservation_stack')
-        reservations = []
-        for reservation_data in reservation_stack:
-            reservation = Reservation(state=Reservation.CONFIRMED, **validated_data, **reservation_data)
-            reservation.save()
-            reservations.append(reservation)
-        
-        instance = ReservationBulk.objects.create(created_by=validated_data['user'])
-        instance.reservations.add(*reservations)
-        
-        return instance
-    
-    def to_representation(self, instance):
-        return ReservationSerializer(
-            context=self.context
-        ).to_representation(instance.reservations.first())
 class HomeMunicipalitySerializer(TranslatedModelSerializer):
     class Meta:
         model = ReservationHomeMunicipalityField
@@ -197,7 +119,9 @@ class HomeMunicipalitySerializer(TranslatedModelSerializer):
             return super().to_internal_value(data)
 
 
-class ReservationSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_api.GeoModelSerializer):
+class ReservationSerializer(ExtraDataMixin, 
+                            ReservationCreateMixin, 
+                            TranslatedModelSerializer, munigeo_api.GeoModelSerializer):
     begin = NullableDateTimeField()
     end = NullableDateTimeField()
     user = UserSerializer(required=False)
@@ -236,53 +160,9 @@ class ReservationSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_a
 
         # set supported and required extra fields
         if resource:
-            cache = self.context.get('reservation_metadata_set_cache')
-            supported = resource.get_supported_reservation_extra_field_names(cache=cache)
-            required = resource.get_required_reservation_extra_field_names(cache=cache)
-
-            # reservations without an order don't require billing fields
-            self.handle_reservation_modify_request(request, resource)
-            order = request.data.get('order')
-
-            begin, end = (request.data.get('begin', None), request.data.get('end', None))
-            if not order or isinstance(order, str) or (order and is_free(get_price(order, begin=begin, end=end))):
-                required = [field for field in required if field not in RESERVATION_BILLING_FIELDS]
-
-            # staff events have less requirements
-            is_staff_event = data.get('staff_event', False)
-
-            if is_staff_event and resource.can_create_staff_event(request.user):
-                required = {'reserver_name', 'event_description'}
-
-            # reservations of type blocked don't require any fields
-            is_blocked_type = data.get('type') == Reservation.TYPE_BLOCKED
-            if is_blocked_type and resource.can_create_special_type_reservation(request.user):
-                required = []
-
-            # we don't need to remove a field here if it isn't supported, as it will be read-only and will be more
-            # easily removed in to_representation()
-            for field_name in supported:
-                self.fields[field_name].read_only = False
-
-            for field_name in required:
-                self.fields[field_name].required = True
+            self.set_supported_and_required_fields(request, resource, data)
 
         self.context.update({'resource': resource})
-
-    def handle_reservation_modify_request(self, request, resource):
-        # handle removing order from data when updating reservation without paying
-        if self.instance and resource.has_products() and 'order' in request.data:
-            state = request.data.get('state')
-            # states where reservation updates can be made
-            if state in (
-                    Reservation.CONFIRMED, Reservation.CANCELLED, Reservation.DENIED,
-                    Reservation.REQUESTED, Reservation.READY_FOR_PAYMENT, Reservation.WAITING_FOR_CASH_PAYMENT):
-                has_staff_perms = resource.is_manager(request.user) or resource.is_admin(request.user)
-                user_can_modify = self.instance.can_modify(request.user)
-                # staff members never pay after reservation creation and their order can be removed safely here
-                # non staff members i.e. clients must include order when state is ready for payment
-                if has_staff_perms or (user_can_modify and state != Reservation.READY_FOR_PAYMENT):
-                    del request.data['order']
 
     def get_required_fields(self):
         return [field_name for field_name in self.fields if self.fields[field_name].required]
@@ -598,6 +478,76 @@ class ReservationSerializer(ExtraDataMixin, TranslatedModelSerializer, munigeo_a
             'can_delete': can_modify_and_delete,
         }
 
+class ReservationStackSerializer(serializers.Serializer):
+    begin = NullableDateTimeField()
+    end = NullableDateTimeField()
+
+class ReservationBulkSerializer(ReservationCreateMixin, serializers.Serializer):
+    resource = serializers.PrimaryKeyRelatedField(queryset=Resource.objects.all())
+    reservation_stack = ReservationStackSerializer(many=True)
+
+    class Meta:
+        fields = ReservationSerializer.Meta.fields + [ 'reservation_stack', 'resource' ]
+
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        for field, value in ReservationSerializer(context=self.context)\
+            .get_fields().items():
+            if field in ('begin', 'end'):
+                continue
+            setattr(self, field, value)
+            self.fields[field] = value
+        data = self.get_initial()
+        request = self.context['request']
+        resource = Resource.objects.get(pk=data['resource'])
+        self.set_supported_and_required_fields(request, resource, data)
+    
+    def validate(self, attrs):
+        attrs = super().validate(attrs)
+
+        resource = attrs['resource']
+        _cattrs = attrs.copy()
+        reservation_stack = _cattrs.pop('reservation_stack')
+        for data in reservation_stack:
+            reservation = Reservation(**_cattrs, **data)
+            reservation.clean()
+            if resource.validate_reservation_period(reservation, reservation.user):
+                return serializers.ValidationError({
+                    'status': False,
+                    'recurring_validation_error': _('Reservation failed. Make sure reservation period is correct.')
+                })
+            if resource.validate_max_reservations_per_user(reservation.user):
+                return  serializers.ValidationError({
+                    'status': False,
+                    'recurring_validation_error': _('Reservation failed. Too many reservations at once.')
+                })
+
+        return attrs
+    
+    def validate_reserver_phone_number(self, value):
+        if value.startswith('+'):
+            if not region_code_for_country_code(phonenumbers.parse(value).country_code):
+                raise ValidationError(dict(reserver_phone_number=_('Invalid country code')))
+        return value
+
+    def create(self, validated_data):
+        reservation_stack = validated_data.pop('reservation_stack')
+        reservations = []
+        for reservation_data in reservation_stack:
+            reservation = Reservation(state=Reservation.CONFIRMED, **validated_data, **reservation_data)
+            reservation.save()
+            reservations.append(reservation)
+        
+        instance = ReservationBulk.objects.create(created_by=validated_data['user'])
+        instance.reservations.add(*reservations)
+        
+        return instance
+    
+    def to_representation(self, instance):
+        return ReservationSerializer(
+            context=self.context
+        ).to_representation(instance.reservations.first())
 
 class UserFilterBackend(filters.BaseFilterBackend):
     """
@@ -884,6 +834,9 @@ class ReservationBulkViewSet(mixins.CreateModelMixin, viewsets.GenericViewSet):
     permission_classes = (
         permissions.IsAuthenticatedOrReadOnly, 
         ReservationPermission, permissions.IsAdminUser,
+    )
+    permission_classes = (
+        permissions.AllowAny, 
     )
     serializer_class = ReservationBulkSerializer
